requirements.txt
aiogram 
detoxify 
torch 
torchvision 
torchaudio 
openai-whisper 
pillow
